manage_projects/requirements.txt
alembic==1.7.4
apipkg==2.1.1
apparmor==3.0.4
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.5.2
attrs==21.4.0
Babel==2.10.3
bcrypt==3.2.1
Beaker==1.11.0
blinker==1.4
Brotli==1.0.9
btrfsutil==5.18.1
CacheControl==0.12.6
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
click==8.1.3
colorama==0.4.4
contextlib2==21.6.0
cryptography==37.0.2
cupshelpers==1.0
decorator==5.1.1
distlib==0.3.4
distro==1.7.0
Django==4.1
djangorestframework==3.13.1
dnspython==2.2.1
docopt==0.6.2
email-validator==1.1.3
extras==1.0.0
fixtures==3.0.0
Flask==2.1.2
Flask-BabelEx==0.9.4
Flask-Compress==1.12
Flask-Gravatar==0.5.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.0.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gssapi==1.7.3
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.1.2
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.8.0
Jinja2==3.1.2
joblib==1.1.0
keyutils==0.6
ldap3==2.9
LibAppArmor==3.0.4
libtorrent==2.0.6
lit==13.0.1.dev0
louis==3.22.0
Mako==1.1.6
manjaro-sdk==0.1
MarkupSafe==2.1.1
more-itertools==8.13.0
msgpack==1.0.3
netsnmp-python==1.0a1
npyscreen==4.10.5
nspektr==0.4.0
numpy==1.22.4
ordered-set==4.0.2
packaging==21.3
pacman-mirrors==4.23.2
pandas==1.4.2
paramiko==2.11.0
passlib==1.7.4
pbr==5.9.0
pep517==0.12.0
Pillow==9.1.1
pluggy==1.0.0
ply==3.11
progress==1.6
psutil==5.9.1
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycairo==1.21.0
pycparser==2.21
pycups==2.0.1
pycurl==7.45.1
PyGObject==3.42.1
PyNaCl==1.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySide6==6.3.1
pysmbc==1.0.23
pytest==7.1.2
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2022.1
PyYAML==6.0
rados==2.0.0
rbd==2.0.0
reportlab==3.6.10
requests==2.27.1
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
scikit-learn==1.1.1
scipy==1.8.1
shiboken6==6.3.1
simplejson==3.17.6
six==1.16.0
speaklater==1.3
SQLAlchemy==1.3.24
sqlparse==0.4.2
sshtunnel==0.4.0
systemd-python==234
team==1.0
testtools==2.5.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
trove-classifiers==2022.6.23
udiskie==2.4.2
urllib3==1.26.9
validate-pyproject==0.9
webencodings==0.5.1
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
